+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/jws/SdJwtSigned.kt
@@ -90,10 +90,10 @@ data class SdJwtSigned(
             hashInput = (listOf(jws.serialize()) + disclosures).joinToString("~", postfix = "~")
         )
 
-        @Deprecated("Use parseThrowing() instead", ReplaceWith("parseThrowing(input)"))
-        fun parse(input: String): SdJwtSigned? = parseThrowing(input).getOrNull()
+        @Deprecated("Use parseCatching() instead", ReplaceWith("parseCatching(input)"))
+        fun parse(input: String): SdJwtSigned? = parseCatching(input).getOrNull()
 
-        fun parseThrowing(input: String): KmmResult<SdJwtSigned> = catching {
+        fun parseCatching(input: String): KmmResult<SdJwtSigned> = catching {
             require(input.contains("~")) { "Could not parse SD-JWT: $input" }
             val stringList = input.replace("[^A-Za-z0-9-_.~]".toRegex(), "").split("~")
             require(stringList.isNotEmpty()) { "Could not parse SD-JWT: $input" }